Troubleshooting Drag and Drop Error
Removed Hermit roaming script as causing hermits to bunch and interfere with functionality.

I also added colliders as children to each character parent - in order to attempt to fix collider issue - however no avail. instead remove this. Had to readd colliders to parents in order for the drag and drop to work. This has now been actioned and i will leave the hermit roaming to a later date. (will ask Aaron for help)