Overhaul of draggable items, dialogue and  camera settings
In order to keep 2.5D working, along with drag and drop functions - I had to change the camera settings, add layers and modify scripts to acknowledge what is draggable when moving collectable items from the UI to the scene whilst still seeing the items ontop of the hermits, and hermits ontop of the background and UI. And have them be read into slider data and dialogue. That now works - Yay!