Reorganized folders
Attempted to fit speech bubble  in prefab. error when making each canvas wordspace. although this is what I want.... It was glitchy and ive run out of ways to fix it without wanting to throw my laptop out the window. It should be a simple fix.... will try again tomorrow after some rest