UI Debugging session
This took hours and was very tedious. However we have a better UI set up now. The headings are still kind of in the way and look different in scale. But the UI us now well out of the way and the icons are larger and more accessible.

Still need to find out why some of the treasured homes wont move when drag and dropped will have to trouble shoot what is being added here.

I also added UI for the tutorial set up